+++ python/paddle/distributed/fleet/layers/mpu/__init__.py
@@ -12,9 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from ...layers.mpu.mp_layers import VocabParallelEmbedding  # noqa: F401
-from ...layers.mpu.mp_layers import ColumnParallelLinear  # noqa: F401
-from ...layers.mpu.mp_layers import RowParallelLinear  # noqa: F401
-from ...layers.mpu.mp_layers import ParallelCrossEntropy  # noqa: F401
+from .mp_layers import VocabParallelEmbedding
+from .mp_layers import ColumnParallelLinear
+from .mp_layers import RowParallelLinear
+from .mp_layers import ParallelCrossEntropy
 
-__all__ = []
+from .random import RNGStatesTracker
+from .random import get_rng_state_tracker
+from .random import model_parallel_random_seed
+from .random import determinate_seed
+from .random import dropout
